+++ examples/postgresql-public-iam/versions.tf
@@ -14,23 +14,6 @@
  * limitations under the License.
  */
 
-output "project_id" {
-  value = var.project_id
-}
-
-output "name" {
-  description = "The name for Cloud SQL instance"
-  value       = module.mysql.instance_name
-}
-
-output "authorized_network" {
-  value = var.mysql_ha_external_ip_range
-}
-
-output "replicas" {
-  value = module.mysql.replicas
-}
-
-output "instances" {
-  value = module.mysql.instances
+terraform {
+  required_version = ">=0.12.6"
 }