+++ modules/mysql/versions.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,10 +15,24 @@
  */
 
 terraform {
-  required_version = ">=0.12.6"
+  required_version = ">= 0.13"
   required_providers {
-    google = "~> 3.22"
-    null   = "~> 2.1"
-    random = "~> 2.2"
+    null = {
+      source  = "hashicorp/null"
+      version = "~> 2.1"
+    }
+    random = {
+      source  = "hashicorp/random"
+      version = "~> 2.2"
+    }
+    google = {
+      source  = "hashicorp/google"
+      version = "~> 3.53"
+    }
   }
+
+  provider_meta "google" {
+    module_name = "blueprints/terraform/terraform-google-sql-db:mysql/v5.1.0"
+  }
+
 }