+++ test/integration/postgresql-public-iam/controls/pg.rb
@@ -12,11 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'json'
-
 project_id = attribute('project_id')
 basename   = attribute('name')
+db_version = "POSTGRES_9_6"
 region = "us-central1"
+tier = "db-f1-micro"
+public_ip_address = attribute('public_ip_address')
 
 activation_policy = "ALWAYS"
 data_disk_size_gb = 10
@@ -26,7 +27,9 @@ pricing_plan = "PER_USE"
 replication_type = "SYNCHRONOUS"
 storage_auto_resize = true
 storage_auto_resize_limit = 0
-tier = "db-custom-2-3840"
+
+cloudsql_iam_user = "dbadmin@goosecorp.org"
+cloudsql_iam_sa = "cloudsql-pg-sa-01@#{project_id}.iam"
 
 describe command("gcloud --project='#{project_id}' sql instances describe #{basename} --format=json") do
   its(:exit_status) { should eq 0 }
@@ -40,7 +43,7 @@ describe command("gcloud --project='#{project_id}' sql instances describe #{base
     end
   end
 
-  describe "mssql_public_database" do
+  describe "postgresql_public_database" do
     it "global settings are valid" do
       expect(data['settings']['activationPolicy']).to eq "#{activation_policy}"
       expect(data['settings']['dataDiskSizeGb']).to eq "#{data_disk_size_gb}"
@@ -58,7 +61,7 @@ describe command("gcloud --project='#{project_id}' sql instances describe #{base
     end
 
     it "database versions is valid" do
-      expect(data['databaseVersion']).to eq 'SQLSERVER_2017_STANDARD'
+      expect(data['databaseVersion']).to eq db_version
     end
 
     it "state is valid" do
@@ -70,13 +73,13 @@ describe command("gcloud --project='#{project_id}' sql instances describe #{base
     end
 
     it "gce zone is valid" do
-      expect(data['gceZone']).to eq "#{region}-a"
+      expect(data['gceZone']).to eq "#{region}-c"
     end
 
     it "location preference is valid" do
       expect(data['settings']['locationPreference']).to include(
-        "kind" => "sql#locationPreference",
-        "zone" => "#{region}-a")
+      "kind" => "sql#locationPreference",
+      "zone" => "#{region}-c")
     end
 
     it "maintenance window is valid" do
@@ -85,6 +88,54 @@ describe command("gcloud --project='#{project_id}' sql instances describe #{base
       "day" => 1,
       "hour" => 23,
       "updateTrack" => "canary")
+    end
+
+    it "database flags are set" do
+      expect(data['settings']['databaseFlags']).to include({
+        "name" => "cloudsql.iam_authentication",
+        "value" => "on"})
+    end
+  end
+
+  describe "Postgres SQL pubic instance" do
+    it "has just one assigned IP address" do
+      expect(data["ipAddresses"].count).to eq(1)
+    end
+
+    it "has expected external IP address" do
+      expect(data["ipAddresses"][0]).to eq(
+        {
+          "type" => "PRIMARY",
+          "ipAddress" => "#{public_ip_address}"
+        }
+      )
+    end
+  end
+end
+
+describe command("gcloud --project='#{project_id}' sql users list --instance #{basename} --format=json") do
+  its(:exit_status) { should eq 0 }
+  its(:stderr) { should eq '' }
+
+  let!(:data) do
+    if subject.exit_status == 0
+      JSON.parse(subject.stdout)
+    else
+      {}
+    end
+  end
+
+  describe "postgresql_public_database" do
+    it "has 1 IAM user" do
+      expect(data.select {|k,v| k['type'] == "CLOUD_IAM_USER"}.size).to eq 1
+      expect(data.select {|k,v| k['name'] == "#{cloudsql_iam_user}"}.size).to eq 1
+    end
+  end
+
+  describe "postgresql_public_database" do
+    it "has 1 IAM service account user" do
+      expect(data.select {|k,v| k['type'] == "CLOUD_IAM_SERVICE_ACCOUNT"}.size).to eq 1
+      expect(data.select {|k,v| k['name'] == "#{cloudsql_iam_sa}"}.size).to eq 1
     end
   end
 end